refactor worker crate into more files
